+++ project-root/backend/src/main/java/com/trip/vo/response/AdminDiaryVO.java
@@ -14,7 +14,7 @@ public class AdminDiaryVO {
     private Long destinationId;
     private Long routeHistoryId;
     private String title;
-    private BigDecimal heatScore;
+    private Long heatScore;
     private BigDecimal ratingScore;
     private Integer ratingCount;
     private String visibility;
@@ -79,11 +79,11 @@ public class AdminDiaryVO {
         this.title = title;
     }
 
-    public BigDecimal getHeatScore() {
+    public Long getHeatScore() {
         return heatScore;
     }
 
-    public void setHeatScore(BigDecimal heatScore) {
+    public void setHeatScore(Long heatScore) {
         this.heatScore = heatScore;
     }
